Optimize size of release
This patch optimizes the size of the released application. It does so by enabling `wasm-opt` to be run on the generated WebAssembly outputs to shave down ~20% of file size and additionally pre-compressing all deployed assets with gzip and brotli. These will be served compressed by GitLab Pages if the client indicates support.