Fix parsing crate responses and indicate if version is yanked
This fixes an issue when parsing crates such as `about` or
`lru-disk-cache` that have their max_stable_version set to null. The fix
changes this field to be option. There is still an issue parsing these
crates because their max_version is set to a version that does not
exist.

Additionally, this indicates in the version dropdown if a crate's
version is yanked by displaying it with a "(yanked)" next to the version
name.