diff-view: introduce a base for folding unchanged lines
Most diff viewers don't show the entire diff since most of it
is useless, but only show the changed lines +/- some context.

Usually the context is 3 lines before and after, with an additional
information in the header in more advanced ones. This can be done
in a subsequent commit, since this is already a lot of code and
strictly better from a usuabily standpoint.