Add basic support for browsing a specific version of a crate
this adds a couple of new path endpoints which can be used to browse the
contents of a specific version of a crate. this is implemented under the hood
as a diff between a crate/version and itself.

When viewing one of these self-diffs, the folding logic is also modified to
default to not folding unchanged hunks, as needing to expand every file while
browsing is inconvenient.