Add basic syntax highlighting support
This patch adds basic syntax highlighting support using the syntect
crate and the default themes & languages supported by that crate.

This unfortunately does increase the size of the wasm blob, and can slow
down display of large documents (especially when building without
optimizations). This appears to mostly be due to the inclusion of the
regex crate and a blob containing language definition regexes.

It may make sense to use more tailored asset sets with only the used
theme & languages in the future to keep down on binary size.