Partially implements mode to diff against a repository contents.
This is not a fully implemented feature, merely a starting point. The
diffing already works and there is partial UI support for it, but it
needs work to properly render a navbar for it and to properly display
error messages.

Also, currently there is a workaround for CORS permission errors.